+++ torch/distributed/_shard/sharded_tensor/__init__.py
@@ -2,9 +2,10 @@
 
 import copy
 import functools
+from typing import List
+
 import torch
 import torch.distributed._shard.sharding_spec as shard_spec
-from typing import List
 
 from .api import (
     _register_sharded_op,
@@ -12,12 +13,10 @@ from .api import (
     ShardedTensor,
     ShardedTensorMetadata,
     TensorProperties,
-    _PartialTensor,
 )
 from .metadata import ShardMetadata  # noqa: F401
+from .partial_tensor import _PartialTensor
 from .utils import load_with_process_group
-import torch.distributed as dist
-from torch.distributed import distributed_c10d
 
 
 def empty(sharding_spec: shard_spec.ShardingSpec,
